[Bugfix][backend]Fixed the get snapshot bug.

+++ paimon-web-server/src/main/java/org/apache/paimon/web/server/data/vo/SnapshotVO.java
@@ -56,9 +56,9 @@ public class SnapshotVO {
 
     private Long changelogRecordCount;
 
-    private Integer addedFileCount;
+    private Long addedFileCount;
 
-    private Integer deletedFileCount;
+    private Long deletedFileCount;
 
     private Long watermark;
 }
